requirements.txt
beautifulsoup4==4.5.3
certifi==2017.1.23
future==0.16.0
peewee==2.8.5
python-telegram-bot==5.3.0
urllib3==1.20
